feat : 요일별, 월별 인구 혼잡도 계산 알고리즘 구현

+++ src/main/java/com/capstone/capstonedesign/repository/DayOfWeekCongestionRepository.java
@@ -3,5 +3,5 @@ package com.capstone.capstonedesign.repository;
 import com.capstone.capstonedesign.domain.entity.congestion.DayOfWeekCongestion;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DayOfWeekCongestionRepository extends JpaRepository<DayOfWeekCongestion, Long> {
+public interface DayOfWeekCongestionRepository extends JpaRepository<DayOfWeekCongestion, Integer> {
 }
